backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
sqlalchemy
psycopg2-binary
pgvector
alembic
bcrypt
openai
cohere
agentmail
svix
pdfplumber
tiktoken
python-dotenv
gunicorn
pytest
thefuzz
python-Levenshtein
